requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
drf-yasg>=1.20,<2.0
deepface>=0.0.76
opencv-python>=4.5.1.48
Pillow>=8.0.0
numpy>=1.19.5
scipy>=1.6.0
shutil
tempfile
